Add correlation of ChIP-seq coverage to motif pairs.

This function first adds ChIP-seq signals along all regions of motif location using the function addCovToGR. Than it calculates the correlation of coverage for each input pair using the function addCovCor. The Pearson correlation coeffiicent is added as new metadata column to the input interactions.

addCor(gi, bwFile, name = "chip", window = 1000, binSize = 1)

Arguments

gi
GInteractions object.

bwFile
File path or connection to BigWig file with ChIP-seq signals.

name
Character indicating the sample name.

window
the window size around the center of ranges in gr.

binSize
size of bins to which the coverage values are combined.