Add correlation of anchor signals to pairs of close genomic regions.

This function adds a vector with correlation values for each input interaction. Only works for input interaction within the given maxDist distance. Note, this function does not work on windows because reading of bigWig files is currently not supported on windows.

addCovCor(gi, datacol = "chip", colname = "cor_chip", maxDist = NULL)

Arguments

gi
A sorted GInteractions object.

datacol
a string matching an annotation column in regions(gi). This column is assumed to hold the same number of values for each interaction as a NumericList.

colname
A string that is used as columnname for the new column in gi.

maxDist
maximal distance of pairs in bp as numeric. If maxDist=NULL, the maximal distance is computed from input interactions gi by max(pairdist(gi)).

Value

A GInteractions similar to gi just with an additional column added.

Examples

if (.Platform$OS.type != "windows") {

  # use internal motif data on chromosome 22
  motifGR <- chromloop::motif.hg19.CTCF.chr22

  # use example bigWig file
  exampleBigWig <- system.file("extdata",
  "GM12878_Stat1.chr22_1-18000000.bigWig", package = "chromloop")

  # add coverage from bigWig file
  motifGR <- addCovToGR(motifGR, exampleBigWig)

  # get all pairs within 1Mb
  gi <- getCisPairs(motifGR, 1e5)

  # compute correaltion of coverge for each pair
  gi <- addCovCor(gi)

  # addCovCor adds a new metadata column:
  mcols(gi)

}

#> DataFrame with 1835 rows and 2 columns
#>           dist   cor_chip
#>      <integer>  <numeric>
#> 1        77347 -0.1371194
#> 2        27104 -0.2117460
#> 3        70206         NA
#> 4        25489  0.7400029
#> 5          291 -0.1884298
#> ...        ...        ...
#> 1831     94722         NA
#> 1832     94096         NA
#> 1833     76231         NA
#> 1834     99787         NA
#> 1835     23556         NA