Optimal cutoff values for logistic regression models.

This dataset contains optimal cutoff scores for the response value of logistic regression models. The cutoff is based on optimal F1-scores. A separate model was trained For each of 121 TF ChIP-seq datasets in human GM12878 cells. The model performance were calculated with Hi-C and ChIA-PET interactions using 10-fold cross-validation.

cutoffByTF

Format

An object of class tbl_df with 121 rows and 3 columns:

TF

Transcription factor name

max_cutoff

The optimal cutoff on the logistic regression response value

max_f1

The optimal f1-score associated to the max_cutoff value