CTCF motif locations in human genome hg19.

A dataset containing the motif hits of the CTCF recognition motif from JASPAR database (MA0139.1) in human genome assembly hg19.

motif.hg19.CTCF

Format

GRanges object with 132739 ranges on positive and negative strand indicated and 5 meta columns:

sequence

genomic DNA sequence

weight

The program scans the input sequences with a position-specific scoring matrix (PSSM) by selecting, at each position, a sequence segment (S) of the same length as the matrix, and assigning a score to this segment. The segment score (weight) is calculated according to the theory developed by Jerry Hertz and Gary Stormo (1999), but with the capability to use Markov chain-based background models, as proposed by Thijs et al. (2001). The weight of a sequence segment (Ws) is calculated as the log-ratio between two probabilities: $$ Ws = log(P(S|M)/P(S|B)) $$ where $$proba_M = P(S|M)$$ is the probability to generate the sequence segment given the matrix. $$proba_B = P(S|B)$$ is the probability to generate the sequence segment given the background model.

Pval

The site-wise P-value estimates the significance of the weight associated to each site. It is computed from the matrix, according to the probabilities described in Staden (1989), Bailey (1998).

ln_Pval

The natural logarithm of the p-value.

sig

The significance, defined as sig = -log_10(P-value)

Source

http://jaspar.genereg.net/html/DOWNLOAD/JASPAR_CORE/pfm/individual/MA0139.1.pfm and http://www.rsat.eu/

Details

The dataset was created by using the tool RSAT matrix-scan (http://www.rsat.eu/) to get all motif hits in the hg19 genome sequence. The PFM motif model for CTCF was downloaded from JASPAR database using this url: http://jaspar.genereg.net/html/DOWNLOAD/JASPAR_CORE/pfm/individual/MA0139.1.pfm The following parameters where used.

-quick

origin start

-bginput

-markov 1

-2str

-uth pval 0.0001

-return pvalue

More information on each parameter can be found on the matrix-scan manual page: http://pedagogix-tagc.univ-mrs.fr/rsat/help.matrix-scan.html

The dataset was than filtered to contain only motif hists with p-value \(\le 10^-6\).